feat: Unificando calculadoras em apenas calculadora principal (funcionando)
