--- src/components/Ipca/Ipca.js
@@ -1,61 +1,0 @@
-const indices = [
-    { ano: 1994, indices: [41.31, 40.27, 42.75, 42.68, 44.03, 47.43, 6.84, 1.86, 1.53, 2.62, 2.81, 1.71] },
-
-    { ano: 1995, indices: [1.7, 1.02, 1.55, 2.43, 2.67, 2.26, 2.36, 0.99, 0.99, 1.41, 1.47, 1.56] },
-
-    { ano: 1996, indices: [1.34, 1.03, 0.35, 1.26, 1.22, 1.19, 1.11, 0.44, 0.15, 0.3, 0.32, 0.47] },
-
-    { ano: 1997, indices: [1.18, 0.5, 0.51, 0.88, 0.41, 0.54, 0.22, -0.02, 0.06, 0.23, 0.17, 0.43] },
-
-    { ano: 1998, indices: [0.71, 0.46, 0.34, 0.24, 0.5, 0.02, -0.12, -0.51, -0.22, 0.02, -0.12, 0.33] },
-
-    { ano: 1999, indices: [0.7, 1.05, 1.1, 0.56, 0.3, 0.19, 1.09, 0.56, 0.31, 1.19, 0.95, 0.6] },
-
-    { ano: 2000, indices: [0.62, 0.13, 0.22, 0.42, 0.01, 0.23, 1.61, 1.31, 0.23, 0.14, 0.32, 0.59] },
-
-    { ano: 2001, indices: [0.57, 0.46, 0.38, 0.58, 0.41, 0.52, 1.33, 0.7, 0.28, 0.83, 0.71, 0.65] },
-
-    { ano: 2002, indices: [0.52, 0.36, 0.6, 0.8, 0.21, 0.42, 1.19, 0.65, 0.72, 1.31, 3.02, 2.1] },
-
-    { ano: 2003, indices: [2.25, 1.57, 1.23, 0.97, 0.61, -0.15, 0.2, 0.34, 0.78, 0.29, 0.34, 0.52] },
-
-    { ano: 2004, indices: [0.76, 0.61, 0.47, 0.37, 0.51, 0.71, 0.91, 0.69, 0.33, 0.44, 0.69, 0.86] },
-
-    { ano: 2005, indices: [0.58, 0.59, 0.61, 0.87, 0.49, -0.02, 0.25, 0.17, 0.35, 0.75, 0.55, 0.36] },
-
-    { ano: 2006, indices: [0.59, 0.41, 0.43, 0.21, 0.1, -0.21, 0.19, 0.05, 0.21, 0.33, 0.31, 0.48] },
-
-    { ano: 2007, indices: [0.44, 0.44, 0.37, 0.25, 0.28, 0.28, 0.24, 0.47, 0.18, 0.3, 0.38, 0.74] },
-
-    { ano: 2008, indices: [0.54, 0.49, 0.48, 0.55, 0.79, 0.74, 0.53, 0.28, 0.26, 0.45, 0.36, 0.28] },
-
-    { ano: 2009, indices: [0.48, 0.55, 0.2, 0.48, 0.47, 0.36, 0.24, 0.15, 0.24, 0.28, 0.41, 0.37] },
-
-    { ano: 2010, indices: [0.75, 0.78, 0.52, 0.57, 0.43, 0.0, 0.01, 0.04, 0.45, 0.75, 0.83, 0.63] },
-
-    { ano: 2011, indices: [0.83, 0.8, 0.79, 0.77, 0.47, 0.15, 0.16, 0.37, 0.53, 0.43, 0.52, 0.5] },
-
-    { ano: 2012, indices: [0.56, 0.45, 0.21, 0.64, 0.36, 0.08, 0.43, 0.41, 0.57, 0.59, 0.6, 0.79] },
-
-    { ano: 2013, indices: [0.86, 0.6, 0.47, 0.55, 0.37, 0.26, 0.03, 0.24, 0.35, 0.57, 0.54, 0.92] },
-
-    { ano: 2014, indices: [0.55, 0.69, 0.92, 0.67, 0.46, 0.4, 0.01, 0.25, 0.57, 0.42, 0.51, 0.78] },
-
-    { ano: 2015, indices: [1.24, 1.22, 1.32, 0.71, 0.74, 0.79, 0.62, 0.22, 0.54, 0.82, 1.01, 0.96] },
-
-    { ano: 2016, indices: [1.27, 0.9, 0.43, 0.61, 0.78, 0.35, 0.52, 0.44, 0.08, 0.26, 0.18, 0.3] },
-
-    { ano: 2017, indices: [0.38, 0.33, 0.25, 0.14, 0.31, -0.23, 0.24, 0.19, 0.16, 0.42, 0.28, 0.44] },
-
-    { ano: 2018, indices: [0.29, 0.32, 0.09, 0.22, 0.4, 1.26, 0.33, -0.09, 0.48, 0.45, -0.21, 0.15] },
-
-    { ano: 2019, indices: [0.32, 0.43, 0.75, 0.57, 0.13, 0.01, 0.19, 0.11, -0.04, 0.1, 0.51, 1.15] },
-
-    { ano: 2020, indices: [0.21, 0.25, 0.07, -0.31, -0.38, 0.26, 0.36, 0.24, 0.64, 0.86, 0.89, 1.35] },
-
-    { ano: 2021, indices: [0.25, 0.86, 0.93, 0.31, 0.83, 0.53, 0.96, 0.87, 1.16, 1.25, 0.95, 0.73] },
-
-    { ano: 2022, indices: [0.54, 1.01, 1.62, 1.06, 0.47, 0.67, -0.68] },
-];
-
-export { indices };
